* examples/oil-inspect.c: (oil_print_impl): use "std." as abbr for standard * examples/uberopt/uberopt.c: (main): fix compile problem on ia64 * examples/work/Makefile.am: glib fixes * liboil/Makefile.am: autobuild a marshalling function * liboil/build_marshal.c: (main), (add_pointer_mask), (print_header), (print_footer): same * liboil/liboilmarshal.c: (_oil_test_marshal_function): the autobuilt source * liboil/liboiltest.c: (oil_test_check_function): use the autobuilt marshalling function * liboil/liboiltest.h: same

+++ examples/uberopt/uberopt.c
@@ -162,7 +162,7 @@ main (int argc, char *argv[])
 
           s = strstr(lines[i],"UBER_INDEX");
           if (s) {
-            g_print("%.*s%d%s\n", (s-lines[i]), lines[i], n_impls, s+10);
+            g_print("%.*s%d%s\n", (int)(s-lines[i]), lines[i], n_impls, s+10);
           } else {
             g_print("%s\n",lines[i]);
           }
@@ -180,7 +180,7 @@ main (int argc, char *argv[])
 
           s = strstr(lines[i],"UBER_INDEX");
           if (s) {
-            g_print("%.*s%d%s\n", (s-lines[i]), lines[i], n_impls, s+10);
+            g_print("%.*s%d%s\n", (int)(s-lines[i]), lines[i], n_impls, s+10);
           } else {
             g_print("%s\n",lines[i]);
           }
